chore: move useNestedGroupExperiments to the hooks folder
